requirements.txt
torch
transformers
flask
flask_cors
pillow
python-dotenv
